メール送信に失敗した場合にエラーメッセージを表示する

+++ lib/Baser/Plugin/Mail/Controller/MailController.php
@@ -360,35 +360,37 @@ class MailController extends MailAppController {
 					}
 
 					// メール送信
-					$this->_sendEmail($sendEmailOptions);
-
-					if (!$this->dbDatas['mailContent']['MailContent']['save_info']) {
-						$fileRecords = [];
-						foreach($this->dbDatas['mailFields'] as $key => $field) {
-							if($field['MailField']['type'] === 'file') {
-								// 削除フラグをセット
-								$fileRecords['MailMessage'] = [
-									$field['MailField']['field_name'] => $this->request->data['MailMessage'][$field['MailField']['field_name']],
-									$field['MailField']['field_name'] . '_delete' => true,
-								];
-								// BcUploadBehavior::deleteFiles() はデータベースのデータを削除する前提となっているため、
-								// Model->data['MailMessage']['field_name'] に、配列ではなく、文字列がセットされている状態を想定しているので状態を模倣する
-								$this->MailMessage->data['MailMessage'][$field['MailField']['field_name']] = $this->request->data['MailMessage'][$field['MailField']['field_name']];
+					if ($this->_sendEmail($sendEmailOptions)) {
+						if (!$this->dbDatas['mailContent']['MailContent']['save_info']) {
+							$fileRecords = [];
+							foreach($this->dbDatas['mailFields'] as $key => $field) {
+								if($field['MailField']['type'] === 'file') {
+									// 削除フラグをセット
+									$fileRecords['MailMessage'] = [
+										$field['MailField']['field_name'] => $this->request->data['MailMessage'][$field['MailField']['field_name']],
+										$field['MailField']['field_name'] . '_delete' => true,
+									];
+									// BcUploadBehavior::deleteFiles() はデータベースのデータを削除する前提となっているため、
+									// Model->data['MailMessage']['field_name'] に、配列ではなく、文字列がセットされている状態を想定しているので状態を模倣する
+									$this->MailMessage->data['MailMessage'][$field['MailField']['field_name']] = $this->request->data['MailMessage'][$field['MailField']['field_name']];
+								}
 							}
+							$this->MailMessage->deleteFiles($fileRecords);
 						}
-						$this->MailMessage->deleteFiles($fileRecords);
+
+						$this->Session->delete('Mail.valid');
+
+						/*** Mail.afterSendEmail ***/
+						$this->dispatchEvent('afterSendEmail', [
+							'data' => $this->request->data
+						]);
+					} else {
+						$this->setMessage(__('エラー : 送信中にエラーが発生しました。しばらくたってから再度送信お願いします。'), true);
+						$this->redirect($this->request->params['Content']['url']);
 					}
-
-					$this->Session->delete('Mail.valid');
-
-					/*** Mail.afterSendEmail ***/
-					$this->dispatchEvent('afterSendEmail', [
-						'data' => $this->request->data
-					]);
 				} else {
-
 					$this->setMessage(__('エラー : 送信中にエラーが発生しました。しばらくたってから再度送信お願いします。'), true);
-					$this->set('sendError', true);
+					$this->redirect($this->request->params['Content']['url']);
 				}
 
 				$this->Session->write('Mail.MailContent', $this->dbDatas['mailContent']);
@@ -540,6 +542,33 @@ class MailController extends MailAppController {
 			}
 		}
 
+		// 管理者に送信
+		if (!empty($adminMail)) {
+			// カンマ区切りで複数設定されていた場合先頭のアドレスをreplayToに利用
+			$userReply = $userMail;
+			if (strpos($userReply, ',') !== false) {
+				list($userReply) = explode(',', $userReply);
+			}
+			$data['other']['mode'] = 'admin';
+			$toAdminOptions = array_merge(
+				[
+					'fromName' => $mailContent['sender_name'],
+					'replyTo' => $userReply,
+					'from' => $fromAdmin,
+					'template' => 'Mail.' . $mailContent['mail_template'],
+					'bcc' => $mailContent['sender_2'],
+					'agentTemplate' => false,
+					'attachments' => $attachments,
+					'additionalParameters' => '-f ' . $fromAdmin,
+				],
+				$options['toAdmin']
+			);
+			$sendResult = $this->sendMail($adminMail, $mailContent['subject_admin'], $data, $toAdminOptions);
+			if (!$sendResult) {
+				return false;
+			}
+		}
+
 		// ユーザーに送信
 		if (!empty($userMail)) {
 			$site = BcSite::findCurrent();
@@ -559,31 +588,13 @@ class MailController extends MailAppController {
 				],
 				$options['toUser']
 			);
-			$this->sendMail($userMail, $mailContent['subject_user'], $data, $toUserOptions);
+			$sendResult = $this->sendMail($userMail, $mailContent['subject_user'], $data, $toUserOptions);
+			if (!$sendResult) {
+				return false;
+			}
 		}
 
-		// 管理者に送信
-		if (!empty($adminMail)) {
-			// カンマ区切りで複数設定されていた場合先頭のアドレスをreplayToに利用
-			if (strpos($userMail, ',') !== false) {
-				list($userMail) = explode(',', $userMail);
-			}
-			$data['other']['mode'] = 'admin';
-			$toAdminOptions = array_merge(
-				[
-					'fromName' => $mailContent['sender_name'],
-					'replyTo' => $userMail,
-					'from' => $fromAdmin,
-					'template' => 'Mail.' . $mailContent['mail_template'],
-					'bcc' => $mailContent['sender_2'],
-					'agentTemplate' => false,
-					'attachments' => $attachments,
-					'additionalParameters' => '-f ' . $fromAdmin,
-				],
-				$options['toAdmin']
-			);
-			$this->sendMail($adminMail, $mailContent['subject_admin'], $data, $toAdminOptions);
-		}
+		return true;
 	}
 	
 /**
